requirements.txt
Django==4.2.16
dj-database-url==2.2.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
supabase==2.9.0
python-dotenv==1.0.1
